Unify all font weights under variables (set in variables.less)
Our site uses a wacky font whose weights are a little non-standard, so we need
to basically dial everything down a notch or two. And if we ever mess with the
font or its configuration again in the future, we'll probably need to dial it
back up (or sideways, etc.). So instead of hardcoding weights calibrated for
more normal fonts, or re-doing the weights in our fonts.css file to try and
predict what "normal" is ahead of time, this commit changes every Less file that
sets a font-weight to use one of the following vars:

* @font-weight-light
* @font-weight-base
* @font-weight-medium
* @font-weight-bold

I've changed all the stock Less to use whichever variable seems to match the
authors' intentions. This doesn't change anything yet on our site, since we're
not really using the more advanced BS widgets, but if we ever do it should make
them look a little better right out of the gate.